MovieTile moved to common folder
